function for solving

+++ src/main/java/com/springex/springrest/repository/FunctionRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FunctionRepository extends JpaRepository<Function, Long> {
+public interface FunctionRepository extends JpaRepository<Function, Integer> {
+
 }
